#1049 P2: filter.rs structural split into compiler/engine/policer
Converts userspace-dp/src/filter.rs (1,419 prod LOC) into a directory
module userspace-dp/src/filter/ with three sub-modules along the
natural conceptual seam the user named in #1049:

- compiler.rs (319 LOC) — Snapshot/AST → typed Filter compiler
  (parse_filter_state, parse_term, parse_address, parse_protocol,
  parse_port_spec, build_port_matcher, build_u8/u6_match_bitmap,
  qualify_filter_key)
- engine.rs (764 LOC) — per-packet filter evaluation
  (evaluate_filter*, evaluate_lo0_filter*, evaluate_interface_*,
  term_matches*, *_affects_tx_selection helpers)
- policer.rs (71 LOC) — token-bucket PolicerState + impl

mod.rs (306 LOC) keeps the shared type vocabulary, constants
(PROTO_TCP/UDP/...), counter helpers (record_filter_counter,
flush_recorded_filter_counters, flush_pending_filter_counter_record,
PendingFilterCounterRecord), and the FilterState/FilterResult/
TxSelectionFilterResult/CachedTxSelectionFilterResult types that all
three sub-modules need.

filter_tests.rs is renamed to filter/tests.rs and the `#[path]` is
updated accordingly.

Pure structural relocation — no logic changes. Function bodies are
byte-for-byte identical. The single `super::` reference in the moved
parse_filter_state (`super::InterfaceSnapshot`) was rewritten to
`crate::InterfaceSnapshot` because `super` from filter/compiler.rs
is `filter`, not the crate root. mod.rs re-imports the snapshot
types it forwards to compiler/tests via `use crate::{...}`.

External callers continue to reach the public API through the
existing `crate::filter::evaluate_filter` etc. paths thanks to
`pub(crate) use compiler::*; pub(crate) use engine::*;
pub(crate) use policer::PolicerState;` re-exports in mod.rs.

cargo test --release: 865 passed, 0 failed.

LOC distribution after:
- filter/mod.rs:        306
- filter/compiler.rs:   319
- filter/engine.rs:     764
- filter/policer.rs:     71
- filter/tests.rs:      863

Each file is well under the 2,000 LOC modularity-discipline threshold,
with clean conceptual boundaries.
